+++ babel.config.js
@@ -1,24 +1,25 @@
 module.exports = {
-  presets: ['module:metro-react-native-babel-preset'],
+  presets: ["module:metro-react-native-babel-preset"],
   plugins: [
     [
-      'module-resolver',
+      "module-resolver",
 
       {
-        root: ['./src'],
-        extensions: ['.ts', '.ios.ts', '.android.ts', '.tsx'],
+        root: ["./src"],
+        extensions: [".ts", ".ios.ts", ".android.ts", ".tsx"],
         alias: {
-          '@assets': './src/assets',
-          '@components': './src/components',
-          '@router': './src/router',
-          '@screens': './src/screens',
-          '@storage': './src/storage',
-          '@utilities': './src/utilities',
-          '@data': './src/data',
-          '@svg': './src/svg',
+          "@assets": "./src/assets",
+          "@components": "./src/components",
+          "@router": "./src/router",
+          "@store": "./src/store",
+          "@screens": "./src/screens",
+          "@storage": "./src/storage",
+          "@utilities": "./src/utilities",
+          "@data": "./src/data",
+          "@svg": "./src/svg",
         },
       },
     ],
-    'react-native-reanimated/plugin', //Added for React Native Animation V2
+    "react-native-reanimated/plugin", //Added for React Native Animation V2
   ],
 };
