+++ src/App.js
@@ -1,8 +1,8 @@
 import React from "react";
 import "react-native-gesture-handler";
 import { NavigationContainer } from "@react-navigation/native";
-import { Router } from "./router";
-import { StoreProvider } from "store/StoreProvider";
+import { Router } from "@router";
+import { StoreProvider } from "@store";
 
 const App = () => {
   return (
