+++ src/store/provider/StoreProvider.type.ts
@@ -1,0 +1,5 @@
+import { ReactNode } from "react";
+
+export type TStoreProvider = {
+  children: ReactNode;
+};
